requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
idna==3.6
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
smartsheet-python-sdk==3.0.2
tzdata==2023.4
urllib3==2.1.0
